Added StoreTrait
- checkAutoIncrementAndAddIfNeed function moved StoreTrait
- Memcached, Buffer fix to use StoreTrait function
